+++ tool/Clean.ps1
@@ -1,5 +1,4 @@
 "Deleting all generated files..."
-Remove-Item bin -ErrorAction Ignore -Force -Recurse
+Remove-Item bin, src/Generated -ErrorAction Ignore -Force -Recurse
 Remove-Item */obj -Force -Recurse
-Remove-Item src/Generated -Force -Recurse
 Remove-Item var/* -Exclude .gitkeep -Force -Recurse