+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version [0.4.0](https://github.com/cedx/html.net/compare/v0.3.0...v0.4.0)
-- Added support for the `on*` event handler attributes.
+- Added support for the `on*` and `tabindex` attributes.
 - Added the `New-BlockquoteElement` cmdlet.
 - Added the `blockquote` alias.
 - Renamed the `New-Doctype` cmdlet to `New-DocumentType`.